- added sms api - opendrive api and box api not yet working

+++ app/src/main/java/com/common/OpenDriveAPI/ODCloseFileUploadRequestInfo.java
@@ -4,9 +4,9 @@ public class ODCloseFileUploadRequestInfo {
 
     public String session_id;
     public String file_id;
-    public int file_size;
+    public long file_size;
     public String temp_location;
-    public int file_time;
+    public long file_time;
     public String access_folder_id;
     public int file_compressed;
     public String file_hash;
